Move the context logic into a custom hook useImage()
To keep the logic
as separate as possible
from the presentation layer (return statement).
