+++ js/src/block-editor/controls/image.js
@@ -16,10 +16,12 @@ import {
 	FormFileUpload,
 	Spinner,
 } from '@wordpress/components';
-import { useSelect } from '@wordpress/data';
-import { mediaUpload } from '@wordpress/editor';
-import { useEffect, useState } from '@wordpress/element';
-import { __, sprintf } from '@wordpress/i18n';
+import { __ } from '@wordpress/i18n';
+
+/**
+ * Internal dependencies
+ */
+import { useImage } from '../hooks';
 
 const allowedTypes = [ 'image' ];
 const defaultImgId = 0;
@@ -27,74 +29,17 @@ const defaultImgId = 0;
 const GcbImageControl = ( props ) => {
 	const { field, getValue, instanceId, onChange } = props;
 	const fieldValue = getValue( props );
-
-	const [ isUploading, setIsUploading ] = useState( false );
-	const [ imageSrc, setImageSrc ] = useState( '' );
-	const [ imageAlt, setImageAlt ] = useState( '' );
-
-	// @ts-ignore: type definition file is missing getMedia().
-	const { getMedia } = useSelect( ( select ) => {
-		return select( 'core' );
-	} );
-
-	useEffect( () => {
-		const newImage = getMedia( fieldValue );
-
-		if ( newImage?.source_url ) { // eslint-disable-line camelcase
-			setImageSrc( newImage.source_url );
-		} else if ( 'string' === typeof newImage ) {
-			// Backwards-compatibility: the fieldValue used to be the URL, not the ID.
-			setImageSrc( newImage );
-		}
-
-		if ( newImage?.alt ) {
-			setImageAlt( newImage.alt );
-		} else if ( newImage?.source_url ) { // eslint-disable-line camelcase
-			setImageAlt(
-				/* translators: %1$s: the image src */
-				sprintf( __( 'This image has no alt attribute, but its src is %1$s', 'genesis-custom-blocks' ), newImage.source_url )
-			);
-		} else {
-			setImageAlt( __( 'This image has no alt attribute', 'genesis-custom-blocks' ) );
-		}
-	}, [ fieldValue, getMedia ] );
-
-	/** @param {Object} image The image to update. */
-	const updateImageSrc = ( image ) => {
-		if ( image?.id ) {
-			onChange( parseInt( image.id ) );
-			setImageSrc( image?.url );
-		}
-	};
-
-	/** @param {Object} image The image that was selected. */
-	const onSelect = ( image ) => {
-		if ( ! image.hasOwnProperty( 'url' ) || ! image.hasOwnProperty( 'id' ) ) {
-			return;
-		}
-		if ( 'blob' === image.url.substr( 0, 4 ) ) {
-			// Still uploading…
-			return;
-		}
-
-		updateImageSrc( image );
-		setIsUploading( false );
-	};
+	const {
+		imageAlt,
+		imageSrc,
+		isUploading,
+		onSelect,
+		setIsUploading,
+		uploadFiles,
+	} = useImage( fieldValue, onChange, allowedTypes );
 
 	const removeImage = () => {
 		onChange( defaultImgId );
-	};
-
-	const uploadFiles = ( files ) => {
-		mediaUpload( {
-			allowedTypes,
-			filesList: files,
-			onFileChange: ( image ) => {
-				onSelect( image[ 0 ] );
-			},
-			maxUploadFileSize: 0,
-			onError: () => {},
-		} );
 	};
 
 	return (
@@ -103,7 +48,7 @@ const GcbImageControl = ( props ) => {
 				? <p className="components-base-control__help">{ field.help }</p>
 				: null
 			}
-			{ imageSrc
+			{ !! imageSrc
 				? (
 					<>
 						<img className="gcb-image__img" src={ imageSrc } alt={ imageAlt } />
