+++ js/src/block-editor/hooks/index.js
@@ -1,0 +1,1 @@
+export { default as useImage } from './useImage';
